Update README.md
lines 542-619 is what I want to work. right now it requests user input, displays the text to be typed, and when the sleep function is done the keyboardinterrupt error is raised. The only change I want made to this is the accuracy and wpm calculations outputted before the error is raised.